refactor: rename layout.tsx to Layout.tsx (PascalCase)
Update import in providers.tsx to match the new filename.
